Move site name into shared
