+++ src/shared/models/site-details/site-name.test.js
